2021.03.14.20:08
Working on Enemy Script
Enemy can now shoot 10 bullets, each every 0.25seconds, with 2seconds reload time. The bullets can follow the player for up to 5seconds.

Enemy shoots or reloads only if the player is in range in front of it. The bullets does no damage and the enemy doesn't have a mover script.